move get_abspos_from_origin_point to better place
