+++ corebehrt/functional/preparation/utils.py
@@ -1,9 +1,5 @@
 from dataclasses import asdict, fields
-from datetime import datetime
-from typing import List, Union
-
-import dask.dataframe as dd
-import pandas as pd
+from typing import List
 
 from corebehrt.modules.preparation.dataset import PatientData
 
@@ -14,27 +10,6 @@ def get_background_length(patients: List[PatientData], vocabulary) -> int:
     example_concepts = patients[0].concepts
     background_length = len(set(example_concepts) & background_tokens)
     return background_length + 2  # +2 for [CLS] and [SEP] tokens
-
-
-def get_abspos_from_origin_point(
-    timestamps: Union[pd.Series, List[datetime]], origin_point: datetime
-) -> Union[pd.Series, List[float]]:
-    """Get the absolute position in hours from the origin point"""
-    if isinstance(timestamps, dd.Series):  # len can cause issues with dask
-        return (timestamps - origin_point).dt.total_seconds() / 60 / 60
-    if isinstance(timestamps, pd.Series):
-        if len(timestamps) == 0:
-            return pd.Series([])
-        return (timestamps - origin_point).dt.total_seconds() / 60 / 60
-    elif isinstance(timestamps, list):
-        return [
-            (timestamp - origin_point).total_seconds() / 60 / 60
-            for timestamp in timestamps
-        ]
-    else:
-        raise TypeError(
-            "Invalid type for timestamps, only pd.Series and list are supported."
-        )
 
 
 def get_non_priority_tokens(vocabulary: dict, low_priority_prefixes: List[str]) -> set:
